+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add handleRemoveFromShoppingList prop to ItemCard component
 - Add removeFromShoppingList prop to InventorySection component
 - Add removeFromShoppingList handler and update item references